Refactoring Existing Types (#11121)
