+++ services/ui-src/src/types/other.ts
@@ -1,0 +1,55 @@
+import React from "react";
+
+// OTHER
+
+export enum PageTypes {
+  STANDARD = "standard",
+  DRAWER = "drawer",
+  MODAL_DRAWER = "modalDrawer",
+  REVIEW_SUBMIT = "reviewSubmit",
+}
+
+// ALERTS
+
+export enum AlertTypes {
+  ERROR = "error",
+  INFO = "info",
+  SUCCESS = "success",
+  WARNING = "warning",
+}
+
+// TIME
+
+export interface DateShape {
+  year: number;
+  month: number;
+  day: number;
+}
+
+export interface TimeShape {
+  hour: number;
+  minute: number;
+  second: number;
+}
+
+// OTHER
+export interface AnyObject {
+  [key: string]: any;
+}
+
+export interface InputChangeEvent extends React.ChangeEvent<HTMLInputElement> {}
+
+export type { IconType } from "react-icons";
+
+export interface TableContentShape {
+  caption?: string;
+  headRow?: string[];
+  bodyRows?: string[][];
+}
+
+export interface CustomHtmlElement {
+  type: string;
+  content: string | any;
+  as?: string;
+  props?: AnyObject;
+}
