+++ services/ui-src/src/types/reports.ts
@@ -1,0 +1,99 @@
+import { AnyObject, CustomHtmlElement, FormJson } from "types";
+
+// REPORT STRUCTURE
+
+export interface ReportJson {
+  id?: string;
+  type?: string;
+  name: string;
+  basePath: string;
+  adminDisabled?: boolean;
+  routes: ReportRoute[];
+  flatRoutes?: ReportRoute[];
+  validationSchema?: AnyObject;
+}
+
+export type ReportRoute = ReportRouteWithForm | ReportRouteWithoutForm;
+
+export interface ReportRouteBase {
+  name: string;
+  path: string;
+  pageType?: string;
+}
+
+export type ReportRouteWithForm =
+  | StandardReportPageShape
+  | DrawerReportPageShape
+  | ModalDrawerReportPageShape;
+
+export interface ReportPageShapeBase extends ReportRouteBase {
+  children?: never;
+  verbiage: ReportPageVerbiage;
+}
+
+export interface StandardReportPageShape extends ReportPageShapeBase {
+  form: FormJson;
+  dashboard?: never;
+  modalForm?: never;
+  drawerForm?: never;
+  entityType?: never;
+}
+
+export interface DrawerReportPageShape extends ReportPageShapeBase {
+  entityType: string;
+  verbiage: DrawerReportPageVerbiage;
+  drawerForm: FormJson;
+  modalForm?: never;
+  form?: never;
+}
+
+export interface ModalDrawerReportPageShape extends ReportPageShapeBase {
+  entityType: string;
+  verbiage: ModalDrawerReportPageVerbiage;
+  modalForm: FormJson;
+  drawerForm: FormJson;
+  form?: never;
+}
+
+export interface ReportRouteWithoutForm extends ReportRouteBase {
+  children?: ReportRoute[];
+  pageType?: string;
+  entityType?: never;
+  verbiage?: never;
+  modalForm?: never;
+  drawerForm?: never;
+  form?: never;
+}
+
+export interface ReportPageVerbiage {
+  intro: {
+    section: string;
+    subsection?: string;
+    spreadsheet?: string;
+    info?: string | CustomHtmlElement[];
+  };
+}
+
+export interface DrawerReportPageVerbiage extends ReportPageVerbiage {
+  dashboardTitle: string;
+  countEntitiesInTitle?: boolean;
+  drawerTitle: string;
+  drawerInfo?: CustomHtmlElement[];
+  missingEntityMessage?: CustomHtmlElement[];
+}
+
+export interface ModalDrawerReportPageVerbiage
+  extends DrawerReportPageVerbiage {
+  addEntityButtonText: string;
+  editEntityButtonText: string;
+  addEditModalAddTitle: string;
+  addEditModalEditTitle: string;
+  addEditModalMessage: string;
+  deleteEntityButtonAltText: string;
+  deleteModalTitle: string;
+  deleteModalConfirmButtonText: string;
+  deleteModalWarning: string;
+  entityUnfinishedMessage: string;
+  enterEntityDetailsButtonText: string;
+  editEntityDetailsButtonText: string;
+}
