+++ services/ui-src/src/utils/auth/UserProvider.tsx
@@ -13,7 +13,7 @@ import config from "config";
 import { initAuthManager, updateTimeout, getExpiration } from "utils";
 import { PRODUCTION_HOST_DOMAIN } from "../../constants";
 // types
-import { MCRUser, UserContextShape, UserRoles } from "types";
+import { MCRUser, UserContextShape, UserRoles } from "types/users";
 
 export const UserContext = createContext<UserContextShape>({
   logout: async () => {},
